Amend all distance indicators to use miles
- Search Results Header
- Search Results Cell
- Venue Summary

+++ lib/Components/list_view_button.dart
@@ -19,7 +19,7 @@ class ListViewButton extends StatefulWidget {
   @required
   final Position userCurrentLocation;
   @required
-  final int searchRadius;
+  final double searchRadius;
   @required
   final String searchTown;
